feat: support collocation solver

+++ tests/test_adjoint_forward.py
@@ -11,8 +11,8 @@ class TestMagnusAdjointForward:
             else:
                 return torch.eye(2).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1)
 
-        y0 = torch.tensor([1.0, 2.0], requires_grad=True)
-        t = torch.tensor([0.0, 0.1, 0.2])
+        y0 = torch.tensor([1.0, 2.0], requires_grad=True, dtype=torch.float64)
+        t = torch.tensor([0.0, 0.1, 0.2], dtype=torch.float64)
         param_keys = []
         param_values = []
 
@@ -22,17 +22,17 @@ class TestMagnusAdjointForward:
 
     def test_magnus_adjoint_forward_nonhomogeneous_magnus(self):
         def system_func(t, params):
-            t_tensor = torch.as_tensor(t)
+            t_tensor = torch.as_tensor(t, dtype=torch.float64)
             if t_tensor.ndim == 0:
-                A = torch.eye(2) * t_tensor
-                g = torch.ones(2) * t_tensor
+                A = torch.eye(2, dtype=torch.float64) * t_tensor
+                g = torch.ones(2, dtype=torch.float64) * t_tensor
             else:
-                A = torch.eye(2).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1)
-                g = torch.ones(t_tensor.shape[0], 2) * t_tensor.unsqueeze(-1)
+                A = torch.eye(2, dtype=torch.float64).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1)
+                g = torch.ones(t_tensor.shape[0], 2, dtype=torch.float64) * t_tensor.unsqueeze(-1)
             return A, g
 
-        y0 = torch.tensor([1.0, 2.0], requires_grad=True)
-        t = torch.tensor([0.0, 0.1, 0.2])
+        y0 = torch.tensor([1.0, 2.0], requires_grad=True, dtype=torch.float64)
+        t = torch.tensor([0.0, 0.1, 0.2], dtype=torch.float64)
         param_keys = []
         param_values = []
 
@@ -42,17 +42,17 @@ class TestMagnusAdjointForward:
 
     def test_magnus_adjoint_forward_nonhomogeneous_glrk(self):
         def system_func(t, params):
-            t_tensor = torch.as_tensor(t)
+            t_tensor = torch.as_tensor(t, dtype=torch.float64)
             if t_tensor.ndim == 0:
-                A = torch.eye(2) * t_tensor
-                g = torch.ones(2) * t_tensor
+                A = torch.eye(2, dtype=torch.float64) * t_tensor
+                g = torch.ones(2, dtype=torch.float64) * t_tensor
             else:
-                A = torch.eye(2).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1)
-                g = torch.ones(t_tensor.shape[0], 2) * t_tensor.unsqueeze(-1)
+                A = torch.eye(2, dtype=torch.float64).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1)
+                g = torch.ones(t_tensor.shape[0], 2, dtype=torch.float64) * t_tensor.unsqueeze(-1)
             return A, g
 
-        y0 = torch.tensor([1.0, 2.0], requires_grad=True)
-        t = torch.tensor([0.0, 0.1, 0.2])
+        y0 = torch.tensor([1.0, 2.0], requires_grad=True, dtype=torch.float64)
+        t = torch.tensor([0.0, 0.1, 0.2], dtype=torch.float64)
         param_keys = []
         param_values = []
 
@@ -61,17 +61,17 @@ class TestMagnusAdjointForward:
         assert y_traj.shape == (3, 2)
 
     def test_magnus_adjoint_forward_with_params(self):
-        p = torch.tensor(0.5, requires_grad=True)
+        p = torch.tensor(0.5, requires_grad=True, dtype=torch.float64)
         def system_func(t, params):
-            t_tensor = torch.as_tensor(t)
+            t_tensor = torch.as_tensor(t, dtype=torch.float64)
             if t_tensor.ndim == 0:
-                A = torch.eye(2) * t_tensor * params['p']
+                A = torch.eye(2, dtype=torch.float64) * t_tensor * params['p']
             else:
-                A = torch.eye(2).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1) * params['p']
+                A = torch.eye(2, dtype=torch.float64).unsqueeze(0).expand(t_tensor.shape[0], -1, -1) * t_tensor.unsqueeze(-1).unsqueeze(-1) * params['p']
             return A
 
-        y0 = torch.tensor([1.0, 2.0], requires_grad=True)
-        t = torch.tensor([0.0, 0.1, 0.2])
+        y0 = torch.tensor([1.0, 2.0], requires_grad=True, dtype=torch.float64)
+        t = torch.tensor([0.0, 0.1, 0.2], dtype=torch.float64)
         param_keys = ['p']
         param_values = [p]
